Fixes arc-batch all-in-one jar
Spring uses internally some META-INF/spring.schemas files to describe
the schema locations of the spring XSD files inside the jar. The content
of the spring.schemas file is not always the same for all Spring
dependencies (spring-beans, spring-context...), so we need to
concatenate them when building the jar to make sure each dependency has
the info it needs.