Clarifies error when failing to load an empty file
requeteFichierBrutalement is expected to returns a select request. When
the file is empty, it used to return an empty string, resulting in an
error later down the line ("org.postgresql.util.PSQLException: ERROR
syntax error near "("").

As a quick fix, it now immediately throws an exception if the file is
empty. It could be improved on later (should empty files fails because
they are empty, or only if no matching rule is found?).